\name{birdCall}
\alias{birdCall}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
How to sound like a real bird
}
\description{
Does one of the three bird calls.
}
\usage{
birdCall(type)
}
\arguments{
  \item{type}{
Object of type integer. Defined values include 1, 2 and 3. Other values return birdCall(1).
}
}
\references{
2. OG Altbau
}
\author{
The pigeon gang
}

\examples{
birdCall()
}
