\name{spritz}
\alias{spritz}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
World's greatest drink!
}
\description{
Tells you how to mix your own Aperol spritz.
}
\usage{
spritz(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{GudeLaune}{
GudeLaune is an object of type "boolean"
}
}
\details{
Prints how to make a good Aperol spritz. If GudeLaune == TRUE, it tells you how to make an even better Aperol spritz
}

\references{
https://www.aperol.com/de/aperol-spritz-ritual
}
\author{
Chris
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
spritz()
}
